Refactor addCohortMemberWS socket handler
